Refactor camera system to use FactorySystem
