+++ src/components/camera/arc-rotate.ts
@@ -1,6 +1,6 @@
 import { Component, ComponentSchema, Types } from 'ecsy';
 import { Vector3 } from '@babylonjs/core/Maths/math.vector';
-import { BabylonTypes } from '../-private/ecsy-types';
+import { BabylonTypes } from '../../-private/ecsy-types';
 
 export default class ArcRotateCamera extends Component<ArcRotateCamera> {
   alpha!: number;
