+++ src/systems/camera.ts
@@ -1,17 +1,11 @@
-import { ComponentConstructor, Entity } from 'ecsy';
-import { ArcRotateCamera, Camera, TransformNode } from '../components';
-import { ArcRotateCamera as BabylonArcRotateCamera } from '@babylonjs/core/Cameras/arcRotateCamera';
+import { Entity } from 'ecsy';
+import { Camera, TransformNode } from '../components';
 import SystemWithCore, { queries } from '../-private/systems/with-core';
 import assert from '../-private/utils/assert';
-import assign from '../-private/utils/assign';
 
 export default class CameraSystem extends SystemWithCore {
   execute(): void {
     super.execute();
-
-    this.queries.arcRotateCamera.added?.forEach((e: Entity) => this.setupArcRotateCamera(e));
-    this.queries.arcRotateCamera.changed?.forEach((e: Entity) => this.updateCamera(e, ArcRotateCamera));
-    this.queries.arcRotateCamera.removed?.forEach((e: Entity) => this.removeCamera(e));
 
     this.queries.camera.added?.forEach((e: Entity) => this.setup(e));
     this.queries.camera.changed?.forEach((e: Entity) => this.update(e));
@@ -19,35 +13,6 @@ export default class CameraSystem extends SystemWithCore {
 
     super.afterExecute();
   }
-
-  // Lifecycle handler for camera builders (ArcRotateCamera etc.)
-
-  setupArcRotateCamera(entity: Entity): void {
-    assert('CameraSystem needs BabylonCoreComponent', this.core);
-
-    const { scene } = this.core;
-    const component = entity.getComponent(ArcRotateCamera)!;
-    const { alpha, beta, radius, target, ...rest } = component;
-    const instance = new BabylonArcRotateCamera(ArcRotateCamera.name, alpha, beta, radius, target, scene, false);
-
-    assign(instance, rest);
-
-    entity.addComponent(Camera, { value: instance });
-  }
-
-  updateCamera(entity: Entity, Component: ComponentConstructor<ArcRotateCamera>): void {
-    const component = entity.getComponent(Component)!;
-    const cameraComponent = entity.getComponent(Camera);
-    assert('No camera instance found', cameraComponent?.value);
-
-    assign(cameraComponent.value, component);
-  }
-
-  removeCamera(entity: Entity): void {
-    entity.removeComponent(Camera);
-  }
-
-  // Lifecycle handler for Camera instance component
 
   setup(entity: Entity): void {
     assert('CameraSystem needs BabylonCoreComponent', this.core);
@@ -97,14 +62,6 @@ export default class CameraSystem extends SystemWithCore {
 
   static queries = {
     ...queries,
-    arcRotateCamera: {
-      components: [ArcRotateCamera],
-      listen: {
-        added: true,
-        changed: true,
-        removed: true,
-      },
-    },
     camera: {
       components: [Camera],
       listen: {
